Simplify the access object trait situation
