+++ packages/data/src/util/index.ts
@@ -1,5 +1,6 @@
 export { CouchAttachmentRecord } from "./CouchAttachmentRecord.js";
 export { FileRef } from "./FileRef.js";
+export { Identified } from "./Identified.js";
 export { ImageRef } from "./ImageRef.js";
 export { MD5 } from "./MD5.js";
 export { Noid } from "./Noid.js";
